cleanup + race reward

+++ src/game/server/entities/character.h
@@ -79,7 +79,7 @@ public:
 	void SaveLoad(int Value);
 	void Tele();
 	void Move(int dir);
-	void Buy(const char *Name, int *Upgrade, int Price, int Click, int Max);
+	void Buy(const char *Name, int *Upgrade, long long unsigned Price, int Click, int Max);
 	int SetActiveUpgrade(int weapon, int upgrade);
 	int MouseEvent(vec2 Pos);
 	int NewPlasma();
